% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_crawl}
\alias{create_crawl}
\title{Create crawl}
\usage{
create_crawl(session_token, account_id, project_id)
}
\arguments{
\item{session_token}{Session token from `create_session_token`.}

\item{account_id}{The ID for the account (can be found with get_account_info)}

\item{project_id}{The ID of the project, from `create_project`}
}
\value{
The crawl's ID number. Also, will define a crawl on deepweb.
}
\description{
Set a project's crawl to start
}
\examples{
crawl_id <- set_crawl_urls('test crawl', sites, session_token, 424)

}

