% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_session_token}
\alias{create_session_token}
\title{Create Session Token}
\usage{
create_session_token(user, api_key)
}
\arguments{
\item{user}{User ID}

\item{api_key}{API key}
}
\value{
A string with the session token
}
\description{
Create the session token needed for other DeepCrawl requests.
Lasts for 6 hours.
}
\examples{

user    <- '42'
api_key <- 'b72OqnAniXHgLAjA8VHCUQ10UQ7EzE-I7eYNLqqWXlSAxYa2JU3AtvRAaIwVDt5kvZJW8Y9tQTjGoOPYb6mIzw'
session_token <- create_session_token(user, api_key)

}

