% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_all_report_data}
\alias{get_all_report_data}
\title{Extract useful report data}
\usage{
get_all_report_data(session_token, account_id, project_id, crawl_id)
}
\arguments{
\item{session_token}{Session token from `create_session_token`.}

\item{account_id}{The ID for the account (can be found with get_account_info)}

\item{project_id}{The ID of the project, from `create_project}

\item{crawl_id}{The ID of the crawl, from `create_crawl`}
}
\value{
A dataframe with one observation for each URL
}
\description{
Get a data frame from the reports object
}
\examples{

get_all_report_data(session_token, 424, 18821, 114538)
}

