% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_crawls_info}
\alias{get_crawls_info}
\title{List all crawls}
\usage{
get_crawls_info(session_token, account_id, project_id)
}
\arguments{
\item{session_token}{Session token from `create_session_token`.}

\item{account_id}{The ID for the account (can be found with get_account_info)}

\item{project_id}{The ID of the project, from `create_project`}
}
\value{
A list with one element for each crawl
}
\description{
List all crawls for a given project
}
\examples{

get_crawls_info(session_token, 424, 18823)
}

