% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_projects_info}
\alias{get_projects_info}
\title{List all projects}
\usage{
get_projects_info(session_token, account_id)
}
\arguments{
\item{session_token}{Session token from `create_session_token`.}

\item{account_id}{The ID for the account (can be found with get_account_info)}
}
\value{
A list with one element for each project
}
\description{
List all projects
}
\examples{

get_projects_info(session_token, 424)
}

