% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_report_data}
\alias{get_report_data}
\title{Get all data from a report}
\usage{
get_report_data(session_token, account_id, project_id, crawl_id, report_id)
}
\arguments{
\item{session_token}{Session token from `create_session_token`.}

\item{account_id}{The ID for the account (can be found with get_account_info)}

\item{project_id}{The ID of the project, from `create_project}

\item{crawl_id}{The ID of the crawl, from `create_crawl`}

\item{report_id}{The ID of the report (can be found within `get_report_info`}
}
\value{
A list with one element for each report
}
\description{
Get the data from a given report
}
\examples{

get_report_data(session_token, 424, 18821, 114538, 42384583)
}

