% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_reports_info}
\alias{get_reports_info}
\title{List all reports}
\usage{
get_reports_info(session_token, account_id, project_id, crawl_id)
}
\arguments{
\item{session_token}{Session token from `create_session_token`.}

\item{account_id}{The ID for the account (can be found with get_account_info)}

\item{project_id}{The ID of the project, from `create_project}

\item{crawl_id}{The ID of the crawl, from `create_crawl`}
}
\value{
A list with one element for each report
}
\description{
List all reports for a given crawl within a project
}
\examples{

get_reports_info(session_token, 424, 18821, 114538)
}

