% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deepcrawl.R
\name{get_reports_temp}
\alias{get_reports_temp}
\title{Get deepcrawl data for a list of sites}
\usage{
get_reports_temp(sites,
  api_key = "b72OqnAniXHgLAjA8VHCUQ10UQ7EzE-I7eYNLqqWXlSAxYa2JU3AtvRAaIwVDt5kvZJW8Y9tQTjGoOPYb6mIzw",
  account_id = 424)
}
\arguments{
\item{sites}{Vector of URLs to get data for}

\item{api_key}{Account API key}

\item{account_id}{Account ID number}
}
\value{
A dataframe with one row per site
}
\description{
Get deepcrawl data for a list of sites
}
\examples{

sites <- c("http://www.independent.co.uk/voices/comment/the-loss-of-scottish-salmon-is-a-cultural-catastrophe-10158356.html",
"http://www.wrs.co.uk/", "http://thecodeofgoodpractice.co.uk/wp-content/uploads/2015/02/cogp_overview.pdf",
"http://www.bbc.co.uk/news/uk-scotland-scotland-business-35643363",
"http://www.bbc.co.uk/news/uk-scotland-scotland-business-35379470",
"http://www.dailymail.co.uk/health/article-205547/Salmon-health-alert.html")

df <- get_deepcrawl(sites)
df

}

